requirements.txt
pandas~=2.2.3
searoute~=1.4.2
fuzzywuzzy~=0.18.0
requests~=2.32.3
python-dotenv~=1.0.1